+++ test/components/AsyncRedirectComponent.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import { Redirect } from "react-router-dom";
 
 class AsyncRedirectComponent extends React.Component {
   static displayName = "AsyncRedirectComponent";
 
-  static async getInitialProps({ req, res, match, history, location, ...ctx }) {
+  static async getInitialProps() {
     return await Promise.resolve({ redirectTo: "/new-location" });
   }
