+++ test/components/NonAsyncRedirectComponent.tsx
